+++ coderd/workspaces_test.go
@@ -1,8 +1,6 @@
 package coderd_test
 
 import (
-	"archive/tar"
-	"bytes"
 	"context"
 	"testing"
 
@@ -34,29 +32,11 @@ func TestWorkspaces(t *testing.T) {
 		})
 		require.NoError(t, err)
 		workspace, err := client.CreateWorkspace(context.Background(), "", coderd.CreateWorkspaceRequest{
-			Name:      "hiii",
+			Name:      "example",
 			ProjectID: project.ID,
 		})
 		require.NoError(t, err)
 		return project, workspace
-	}
-
-	setupProjectVersion := func(t *testing.T, client *codersdk.Client, user coderd.CreateInitialUserRequest, project coderd.Project) coderd.ProjectHistory {
-		var buffer bytes.Buffer
-		writer := tar.NewWriter(&buffer)
-		err := writer.WriteHeader(&tar.Header{
-			Name: "file",
-			Size: 1 << 10,
-		})
-		require.NoError(t, err)
-		_, err = writer.Write(make([]byte, 1<<10))
-		require.NoError(t, err)
-		projectHistory, err := client.CreateProjectHistory(context.Background(), user.Organization, project.Name, coderd.CreateProjectVersionRequest{
-			StorageMethod: database.ProjectStorageMethodInlineArchive,
-			StorageSource: buffer.Bytes(),
-		})
-		require.NoError(t, err)
-		return projectHistory
 	}
 
 	t.Run("List", func(t *testing.T) {
@@ -132,12 +112,12 @@ func TestWorkspaces(t *testing.T) {
 		_, err = server.Client.CreateUser(context.Background(), coderd.CreateUserRequest{
 			Email:    "hello@ok.io",
 			Username: "example",
-			Password: "wowowow",
+			Password: "password",
 		})
 		require.NoError(t, err)
 		token, err := server.Client.LoginWithPassword(context.Background(), coderd.LoginWithPasswordRequest{
 			Email:    "hello@ok.io",
-			Password: "wowowow",
+			Password: "password",
 		})
 		require.NoError(t, err)
 		err = server.Client.SetSessionToken(token.SessionToken)
@@ -168,88 +148,5 @@ func TestWorkspaces(t *testing.T) {
 		_, workspace := setupProjectAndWorkspace(t, server.Client, user)
 		_, err := server.Client.Workspace(context.Background(), "", workspace.Name)
 		require.NoError(t, err)
-	})
-
-	t.Run("AllHistory", func(t *testing.T) {
-		t.Parallel()
-		server := coderdtest.New(t)
-		user := server.RandomInitialUser(t)
-		project, workspace := setupProjectAndWorkspace(t, server.Client, user)
-		history, err := server.Client.WorkspaceHistory(context.Background(), "", workspace.Name)
-		require.NoError(t, err)
-		require.Len(t, history, 0)
-		projectVersion := setupProjectVersion(t, server.Client, user, project)
-		_, err = server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: projectVersion.ID,
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.NoError(t, err)
-		history, err = server.Client.WorkspaceHistory(context.Background(), "", workspace.Name)
-		require.NoError(t, err)
-		require.Len(t, history, 1)
-	})
-
-	t.Run("LatestHistory", func(t *testing.T) {
-		t.Parallel()
-		server := coderdtest.New(t)
-		user := server.RandomInitialUser(t)
-		project, workspace := setupProjectAndWorkspace(t, server.Client, user)
-		_, err := server.Client.LatestWorkspaceHistory(context.Background(), "", workspace.Name)
-		require.Error(t, err)
-		projectVersion := setupProjectVersion(t, server.Client, user, project)
-		_, err = server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: projectVersion.ID,
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.NoError(t, err)
-		_, err = server.Client.LatestWorkspaceHistory(context.Background(), "", workspace.Name)
-		require.NoError(t, err)
-	})
-
-	t.Run("CreateHistory", func(t *testing.T) {
-		t.Parallel()
-		server := coderdtest.New(t)
-		user := server.RandomInitialUser(t)
-		project, workspace := setupProjectAndWorkspace(t, server.Client, user)
-		projectHistory := setupProjectVersion(t, server.Client, user, project)
-
-		_, err := server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: projectHistory.ID,
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.NoError(t, err)
-	})
-
-	t.Run("CreateHistoryAlreadyInProgress", func(t *testing.T) {
-		t.Parallel()
-		server := coderdtest.New(t)
-		user := server.RandomInitialUser(t)
-		project, workspace := setupProjectAndWorkspace(t, server.Client, user)
-		projectHistory := setupProjectVersion(t, server.Client, user, project)
-
-		_, err := server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: projectHistory.ID,
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.NoError(t, err)
-
-		_, err = server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: projectHistory.ID,
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.Error(t, err)
-	})
-
-	t.Run("CreateHistoryInvalidProjectVersion", func(t *testing.T) {
-		t.Parallel()
-		server := coderdtest.New(t)
-		user := server.RandomInitialUser(t)
-		_, workspace := setupProjectAndWorkspace(t, server.Client, user)
-
-		_, err := server.Client.CreateWorkspaceHistory(context.Background(), "", workspace.Name, coderd.CreateWorkspaceHistoryRequest{
-			ProjectHistoryID: uuid.New(),
-			Transition:       database.WorkspaceTransitionCreate,
-		})
-		require.Error(t, err)
 	})
 }
